Week2Push
Here is my week 2 push for Project 1! I added the face of a yogi, and I added a sine curve to portray a wave that flows into an ocean. I also started working on creating an array of flowers and started creating a heart shape, which I will later animate to "breathe" or "beat" in the chest of the yogi.